The connection at redux changed from Books to BookWrapper
